update proftp
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@17876 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/proftpd/modules/mod_ctrls.c
@@ -3,7 +3,7 @@
  *          server, as well as several utility functions for other Controls
  *          modules
  *
- * Copyright (c) 2000-2009 TJ Saunders
+ * Copyright (c) 2000-2011 TJ Saunders
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,7 +17,7 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307, USA.
+ * Foundation, Inc., 51 Franklin Street, Suite 500, Boston, MA 02110-1335, USA.
  *
  * As a special exemption, TJ Saunders and other respective copyright holders
  * give permission to link this program with OpenSSL, and distribute the
@@ -27,7 +27,7 @@
  * This is mod_ctrls, contrib software for proftpd 1.2 and above.
  * For more information contact TJ Saunders <tj@castaglia.org>.
  *
- * $Id: mod_ctrls.c,v 1.45 2009/07/17 01:16:02 castaglia Exp $
+ * $Id: mod_ctrls.c,v 1.50 2011/07/31 22:07:03 castaglia Exp $
  */
 
 #include "conf.h"
@@ -78,6 +78,8 @@ static unsigned int cl_maxlistlen = 5;
 static ctrls_acl_t ctrls_sock_acl;
 
 static unsigned char ctrls_engine = TRUE;
+
+#define CTRLS_LISTEN_FL_REMOVE_SOCKET	0x0001
 
 /* Necessary prototypes */
 static int ctrls_setblock(int sockfd);
@@ -426,7 +428,7 @@ static int ctrls_cls_write(void) {
 }
 
 /* Create a listening local socket */
-static int ctrls_listen(const char *sock_file) {
+static int ctrls_listen(const char *sock_file, int flags) {
   int sockfd = -1, len = 0;
   struct sockaddr_un sock;
 #if !defined(SO_PEERCRED) && !defined(HAVE_GETPEEREID) && \
@@ -473,10 +475,12 @@ static int ctrls_listen(const char *sock_file) {
     }
   }
 
-  /* Make sure the path to which we want to bind this socket doesn't already
-   * exist.
-   */
-  unlink(sock_file);
+  if (flags & CTRLS_LISTEN_FL_REMOVE_SOCKET) {
+    /* Make sure the path to which we want to bind this socket doesn't already
+     * exist.
+     */
+    unlink(sock_file);
+  }
 
   /* Fill in the socket structure fields */
   memset(&sock, 0, sizeof(sock));
@@ -493,11 +497,14 @@ static int ctrls_listen(const char *sock_file) {
 
     pr_signals_unblock();
     (void) close(sockfd);
+
     errno = xerrno;
     pr_ctrls_log(MOD_CTRLS_VERSION,
       "error: unable to bind to local socket: %s", strerror(xerrno));
     pr_trace_msg(trace_channel, 1, "unable to bind to local socket: %s",
       strerror(xerrno));
+
+    errno = xerrno;
     return -1;
   }
 
@@ -507,11 +514,15 @@ static int ctrls_listen(const char *sock_file) {
 
     pr_signals_unblock();
     (void) close(sockfd);
+
     errno = xerrno;
     pr_ctrls_log(MOD_CTRLS_VERSION,
-      "error: unable to listen on local socket: %s", strerror(xerrno));
-    pr_trace_msg(trace_channel, 1, "unable to listen on local socket: %s",
+      "error: unable to listen on local socket '%s': %s", sock.sun_path,
       strerror(xerrno));
+    pr_trace_msg(trace_channel, 1, "unable to listen on local socket '%s': %s",
+      sock.sun_path, strerror(xerrno));
+
+    errno = xerrno;
     return -1;
   }
 
@@ -529,11 +540,14 @@ static int ctrls_listen(const char *sock_file) {
 
     pr_signals_unblock();
     (void) close(sockfd);
+
     errno = xerrno;
     pr_ctrls_log(MOD_CTRLS_VERSION,
       "error: unable to chmod local socket: %s", strerror(xerrno));
     pr_trace_msg(trace_channel, 1, "unable to chmod local socket: %s",
       strerror(xerrno));
+
+    errno = xerrno;
     return -1;
   }
 
@@ -849,9 +863,9 @@ static int ctrls_handle_rmctrl(pr_ctrls_t *ctrl, int reqargc,
   /* The three controls added by this module _cannot_ be removed (at least
    * not via this control handler).
    */
-  if (strcmp(reqargv[0], "insctrl") == 0 ||
-      strcmp(reqargv[0], "lsctrl") == 0 ||
-      strcmp(reqargv[0], "rmctrl") == 0) {
+  if (strncmp(reqargv[0], "insctrl", 8) == 0 ||
+      strncmp(reqargv[0], "lsctrl", 7) == 0 ||
+      strncmp(reqargv[0], "rmctrl", 7) == 0) {
     pr_ctrls_add_response(ctrl, "'%s' control cannot be removed", reqargv[0]);
     return -1;
   }
@@ -864,18 +878,21 @@ static int ctrls_handle_rmctrl(pr_ctrls_t *ctrl, int reqargc,
 
   if (pr_set_registered_actions(m, reqargv[0], FALSE,
       PR_CTRLS_ACT_DISABLED) < 0) {
+    int xerrno = errno;
 
-    if (errno == ENOENT)
+    if (xerrno == ENOENT) {
       pr_ctrls_add_response(ctrl, "no such control: '%s'", reqargv[0]);
-    else
+
+    } else {
       pr_ctrls_add_response(ctrl, "unable to disable '%s': %s", reqargv[0],
-        strerror(errno));
+        strerror(xerrno));
+    }
 
   } else {
-    if (strcmp(reqargv[0], "all") != 0) 
+    if (strncmp(reqargv[0], "all", 4) != 0) {
       pr_ctrls_add_response(ctrl, "'%s' control disabled", reqargv[0]);
 
-    else {
+    } else {
 
       /* If all actions have been disabled, stop listening on the local
        * socket, and turn off this module's engine.
@@ -914,14 +931,16 @@ MODRET set_ctrlsacls(cmd_rec *cmd) {
   actions = ctrls_parse_acl(cmd->tmp_pool, cmd->argv[1]);
 
   /* Check the second parameter to make sure it is "allow" or "deny" */
-  if (strcmp(cmd->argv[2], "allow") != 0 &&
-      strcmp(cmd->argv[2], "deny") != 0)
+  if (strncmp(cmd->argv[2], "allow", 6) != 0 &&
+      strncmp(cmd->argv[2], "deny", 5) != 0) {
     CONF_ERROR(cmd, "second parameter must be 'allow' or 'deny'");
+  }
 
   /* Check the third parameter to make sure it is "user" or "group" */
-  if (strcmp(cmd->argv[3], "user") != 0 &&
-      strcmp(cmd->argv[3], "group") != 0)
+  if (strncmp(cmd->argv[3], "user", 5) != 0 &&
+      strncmp(cmd->argv[3], "group", 6) != 0) {
     CONF_ERROR(cmd, "third parameter must be 'user' or 'group'");
+  }
 
   bad_action = pr_ctrls_set_module_acls(ctrls_acttab, ctrls_pool, actions,
     cmd->argv[2], cmd->argv[3], cmd->argv[4]);
@@ -1052,16 +1071,17 @@ MODRET set_ctrlssocketacl(cmd_rec *cmd) {
   pr_ctrls_init_acl(&ctrls_sock_acl);
 
   /* Check the first argument to make sure it either "allow" or "deny" */
-  if (strcmp(cmd->argv[1], "allow") != 0 &&
-      strcmp(cmd->argv[1], "deny") != 0)
+  if (strncmp(cmd->argv[1], "allow", 6) != 0 &&
+      strncmp(cmd->argv[1], "deny", 5) != 0) {
     CONF_ERROR(cmd, "first parameter must be either 'allow' or 'deny'");
+  }
 
   /* Check the second argument to see how to handle the directive */
-  if (strcmp(cmd->argv[2], "user") == 0) {
+  if (strncmp(cmd->argv[2], "user", 5) == 0) {
     pr_ctrls_set_user_acl(ctrls_pool, &ctrls_sock_acl.acl_usrs, cmd->argv[1],
       cmd->argv[3]);
  
-  } else if (strcmp(cmd->argv[2], "group") == 0) {
+  } else if (strncmp(cmd->argv[2], "group", 6) == 0) {
     pr_ctrls_set_group_acl(ctrls_pool, &ctrls_sock_acl.acl_grps, cmd->argv[1],
       cmd->argv[3]);
 
@@ -1112,7 +1132,7 @@ MODRET set_ctrlssocketowner(cmd_rec *cmd) {
 /* Event handlers
  */
 
-static void ctrls_exit_ev(const void *event_data, void *user_data) {
+static void ctrls_shutdown_ev(const void *event_data, void *user_data) {
   if (!is_master || !ctrls_engine)
     return;
 
@@ -1127,22 +1147,6 @@ static void ctrls_exit_ev(const void *event_data, void *user_data) {
   }
 
   return;
-}
-
-static void ctrls_postparse_ev(const void *event_data, void *user_data) {
-
-  if (!ctrls_engine) {
-    return;
-  }
-
-  /* Start listening on the ctrl socket */
-  PRIVS_ROOT
-  ctrls_sockfd = ctrls_listen(ctrls_sock_file);
-  PRIVS_RELINQUISH
-  if (ctrls_sockfd < 0) {
-    pr_log_pri(PR_LOG_NOTICE, "notice: unable to listen to local socket: %s",
-      strerror(errno));
-  }
 }
 
 static void ctrls_restart_ev(const void *event_data, void *user_data) {
@@ -1197,6 +1201,14 @@ static void ctrls_restart_ev(const void *event_data, void *user_data) {
 }
 
 static void ctrls_startup_ev(const void *event_data, void *user_data) {
+  if (!ctrls_engine) {
+    return;
+  }
+
+  /* Start listening on the ctrl socket */
+  PRIVS_ROOT
+  ctrls_sockfd = ctrls_listen(ctrls_sock_file, CTRLS_LISTEN_FL_REMOVE_SOCKET);
+  PRIVS_RELINQUISH
 
   /* Start a timer for the checking/processing of the ctrl socket.  */
   pr_timer_remove(CTRLS_TIMER_ID, &ctrls_module);
@@ -1233,15 +1245,10 @@ static int ctrls_init(void) {
   ctrls_sock_acl.acl_usrs.allow = ctrls_sock_acl.acl_grps.allow = FALSE;
 
   /* Start listening on the ctrl socket */
-  ctrls_sockfd = ctrls_listen(ctrls_sock_file);
-  if (ctrls_sockfd < 0) {
-    pr_log_pri(PR_LOG_NOTICE, "notice: unable to listen to local socket: %s",
-      strerror(errno));
-  }
+  ctrls_sockfd = ctrls_listen(ctrls_sock_file, 0);
 
-  pr_event_register(&ctrls_module, "core.exit", ctrls_exit_ev, NULL);
-  pr_event_register(&ctrls_module, "core.postparse", ctrls_postparse_ev, NULL);
   pr_event_register(&ctrls_module, "core.restart", ctrls_restart_ev, NULL);
+  pr_event_register(&ctrls_module, "core.shutdown", ctrls_shutdown_ev, NULL);
   pr_event_register(&ctrls_module, "core.startup", ctrls_startup_ev, NULL);
 
   return 0;
@@ -1253,7 +1260,6 @@ static int ctrls_sess_init(void) {
   ctrls_engine = FALSE;
   pr_timer_remove(CTRLS_TIMER_ID, &ctrls_module);
 
-  pr_event_unregister(&ctrls_module, "core.exit", ctrls_exit_ev);
   pr_event_unregister(&ctrls_module, "core.restart", ctrls_restart_ev);
 
   /* Close the inherited socket */
